requirements.txt
# Core Dependencies
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
numpy>=1.21.0
scipy>=1.7.0

# Image Processing and Metrics
scikit-image>=0.18.0
matplotlib>=3.4.0
Pillow>=8.0.0

# Configuration and Logging
PyYAML>=5.4.0
wandb>=0.12.0  # Optional, for experiment tracking

# Utilities
tqdm>=4.62.0
pathlib2>=2.3.0  # For Python < 3.4 compatibility

# Development and Testing (Optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0
isort>=5.9.0

# Medical Imaging (Optional)
# SimpleITK>=2.1.0  # Uncomment if working with medical images
# nibabel>=3.2.0    # Uncomment if working with NIfTI files
